Fix bug when map wasn't centered on last added way point
WayPoint - Ya Maps API object representing one point of the MultiRoute.
@see tech.yandex.ru/maps/doc/jsapi/2.1/ref/reference/multiRouter.WayPoin
t-docpage/

LocationPoint - this project's term determining object in locations
array.

Because of the internal structure of MultiRoute (points probably are
stored unodered with some property specifing their positions) if the
referencePoints (aka wayPoints) array is updated with the same data but
in different order the last item will be returned as before with no
updates.

So I have to filter through array searching for match between wayPoint
and locationPoint being the last in locations array.