Add half-working sync to update locations after wayPoint was dragged
The implementation is called half-working because it's based on
address (words) not coordinates.
Because of that after the way point was dragged, it can randomly
jump as Yandex Map may match a slightly different point for that
address (read a set of words), not the where the user placed it.